Create header footer container
